Made unsigned int as unsigned long long int

+++ lib/Trajectory.cpp
@@ -590,7 +590,7 @@ unsigned int Trajectory::getLastFrame(){
   return lastFrame;
 }
 
-unsigned int Trajectory::getLastPos(){
+unsigned long long int Trajectory::getLastPos(){
 	return lastPos;
 }
 
@@ -714,7 +714,7 @@ void Trajectory::setLastFrame(const unsigned int &frame){
   lastFrame=frame;
 }
 
-void Trajectory::setLastPos(const unsigned int &pos){
+void Trajectory::setLastPos(const unsigned long long int &pos){
 	lastPos=pos;
 }
